move new loading state test to its own file
